try and fix ant installation on mac

+++ pyflutterinstall/install/ant_sdk.py
@@ -10,6 +10,7 @@ import sys
 from pathlib import Path
 
 from download import download  # type: ignore
+from shellexecute import execute  # type: ignore
 
 from pyflutterinstall.resources import (
     ANT_DIR,
@@ -23,20 +24,26 @@ from pyflutterinstall.util import make_title
 
 def install_ant_sdk() -> int:
     make_title("Installing Ant SDK")
-    print(f"Install Ant from {ANT_SDK_DOWNLOAD} to {INSTALL_DIR}")
-    ant_sdk_sip = Path(
-        download(ANT_SDK_DOWNLOAD, DOWNLOAD_DIR / os.path.basename(ANT_SDK_DOWNLOAD))
-    )
-    if os.path.exists(ANT_DIR):
-        print(f"Removing existing Ant SDK at {ANT_DIR}")
-        shutil.rmtree(ANT_DIR)
-    print(f"Unpacking {ant_sdk_sip} to {ANT_DIR}")
-    shutil.unpack_archive(ant_sdk_sip, ANT_DIR)
-    base_ant_dir = ANT_DIR / os.listdir(ANT_DIR)[0]
-    print(base_ant_dir)
-    ant_bin_dir = base_ant_dir / "bin"
-    add_env_path(ant_bin_dir)
-    # check that ant is in the path
+    if sys.platform == "win32":
+        print(f"Install Ant from {ANT_SDK_DOWNLOAD} to {INSTALL_DIR}")
+        ant_sdk_sip = Path(
+            download(
+                ANT_SDK_DOWNLOAD, DOWNLOAD_DIR / os.path.basename(ANT_SDK_DOWNLOAD)
+            )
+        )
+        if os.path.exists(ANT_DIR):
+            print(f"Removing existing Ant SDK at {ANT_DIR}")
+            shutil.rmtree(ANT_DIR)
+        print(f"Unpacking {ant_sdk_sip} to {ANT_DIR}")
+        shutil.unpack_archive(ant_sdk_sip, ANT_DIR)
+        base_ant_dir = ANT_DIR / os.listdir(ANT_DIR)[0]
+        print(base_ant_dir)
+        ant_bin_dir = base_ant_dir / "bin"
+        add_env_path(ant_bin_dir)
+        # check that ant is in the path
+    if sys.platform == "darwin":
+        execute("brew install ant")
+        return 1
     assert shutil.which("ant"), "ant not found in ant bin dir"
     print(f"Ant SDK installed: {base_ant_dir}\n")
     return 0
